+++ src/muleengine/mesh_creator/rectangle3d.rs
@@ -58,14 +58,14 @@ pub fn create(width: f32, height: f32, depth: f32) -> Mesh {
     );
 
     mesh.add_face(
-        mesh.number_of_vertices() - 4,
-        mesh.number_of_vertices() - 3,
-        mesh.number_of_vertices() - 2,
+        (mesh.number_of_vertices() - 4) as u32,
+        (mesh.number_of_vertices() - 3) as u32,
+        (mesh.number_of_vertices() - 2) as u32,
     );
     mesh.add_face(
-        mesh.number_of_vertices() - 4,
-        mesh.number_of_vertices() - 2,
-        mesh.number_of_vertices() - 1,
+        (mesh.number_of_vertices() - 4) as u32,
+        (mesh.number_of_vertices() - 2) as u32,
+        (mesh.number_of_vertices() - 1) as u32,
     );
 
     // right
@@ -115,14 +115,14 @@ pub fn create(width: f32, height: f32, depth: f32) -> Mesh {
     );
 
     mesh.add_face(
-        mesh.number_of_vertices() - 4,
-        mesh.number_of_vertices() - 3,
-        mesh.number_of_vertices() - 2,
+        (mesh.number_of_vertices() - 4) as u32,
+        (mesh.number_of_vertices() - 3) as u32,
+        (mesh.number_of_vertices() - 2) as u32,
     );
     mesh.add_face(
-        mesh.number_of_vertices() - 4,
-        mesh.number_of_vertices() - 2,
-        mesh.number_of_vertices() - 1,
+        (mesh.number_of_vertices() - 4) as u32,
+        (mesh.number_of_vertices() - 2) as u32,
+        (mesh.number_of_vertices() - 1) as u32,
     );
 
     // top
@@ -172,14 +172,14 @@ pub fn create(width: f32, height: f32, depth: f32) -> Mesh {
     );
 
     mesh.add_face(
-        mesh.number_of_vertices() - 4,
-        mesh.number_of_vertices() - 3,
-        mesh.number_of_vertices() - 2,
+        (mesh.number_of_vertices() - 4) as u32,
+        (mesh.number_of_vertices() - 3) as u32,
+        (mesh.number_of_vertices() - 2) as u32,
     );
     mesh.add_face(
-        mesh.number_of_vertices() - 4,
-        mesh.number_of_vertices() - 2,
-        mesh.number_of_vertices() - 1,
+        (mesh.number_of_vertices() - 4) as u32,
+        (mesh.number_of_vertices() - 2) as u32,
+        (mesh.number_of_vertices() - 1) as u32,
     );
 
     // bottom
@@ -229,14 +229,14 @@ pub fn create(width: f32, height: f32, depth: f32) -> Mesh {
     );
 
     mesh.add_face(
-        mesh.number_of_vertices() - 4,
-        mesh.number_of_vertices() - 3,
-        mesh.number_of_vertices() - 2,
+        (mesh.number_of_vertices() - 4) as u32,
+        (mesh.number_of_vertices() - 3) as u32,
+        (mesh.number_of_vertices() - 2) as u32,
     );
     mesh.add_face(
-        mesh.number_of_vertices() - 4,
-        mesh.number_of_vertices() - 2,
-        mesh.number_of_vertices() - 1,
+        (mesh.number_of_vertices() - 4) as u32,
+        (mesh.number_of_vertices() - 2) as u32,
+        (mesh.number_of_vertices() - 1) as u32,
     );
 
     // front
@@ -286,14 +286,14 @@ pub fn create(width: f32, height: f32, depth: f32) -> Mesh {
     );
 
     mesh.add_face(
-        mesh.number_of_vertices() - 4,
-        mesh.number_of_vertices() - 3,
-        mesh.number_of_vertices() - 2,
+        (mesh.number_of_vertices() - 4) as u32,
+        (mesh.number_of_vertices() - 3) as u32,
+        (mesh.number_of_vertices() - 2) as u32,
     );
     mesh.add_face(
-        mesh.number_of_vertices() - 4,
-        mesh.number_of_vertices() - 2,
-        mesh.number_of_vertices() - 1,
+        (mesh.number_of_vertices() - 4) as u32,
+        (mesh.number_of_vertices() - 2) as u32,
+        (mesh.number_of_vertices() - 1) as u32,
     );
 
     // back
@@ -343,14 +343,14 @@ pub fn create(width: f32, height: f32, depth: f32) -> Mesh {
     );
 
     mesh.add_face(
-        mesh.number_of_vertices() - 4,
-        mesh.number_of_vertices() - 3,
-        mesh.number_of_vertices() - 2,
+        (mesh.number_of_vertices() - 4) as u32,
+        (mesh.number_of_vertices() - 3) as u32,
+        (mesh.number_of_vertices() - 2) as u32,
     );
     mesh.add_face(
-        mesh.number_of_vertices() - 4,
-        mesh.number_of_vertices() - 2,
-        mesh.number_of_vertices() - 1,
+        (mesh.number_of_vertices() - 4) as u32,
+        (mesh.number_of_vertices() - 2) as u32,
+        (mesh.number_of_vertices() - 1) as u32,
     );
 
     mesh.compute_tangents(0);
